chore(starknet_integration_tests): use iterator instead of vec push
chore(starknet_integration_tests): prepend sequencer part to test output

chore(starknet_integration_tests): awaiting aliveness concurrently

refactor(starknet_integration_tests): move fns to sequencer manager mod

refactor(starknet_integration_tests): move more fns to sequencer manager mod

commit-id:eff2ca71
