+++ crates/starknet_sequencer_infra/src/test_utils.rs
@@ -3,7 +3,7 @@ use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use tokio::net::TcpListener;
 
 const PORTS_PER_INSTANCE: u16 = 60;
-const MAX_NUMBER_OF_INSTANCES_PER_TEST: u16 = 10;
+pub const MAX_NUMBER_OF_INSTANCES_PER_TEST: u16 = 10;
 const MAX_NUMBER_OF_TESTS: u16 = 10;
 const BASE_PORT: u16 = 55000;
